requirements.txt
gradio>=4.0
openai>=1.0
anthropic>=0.40.0
pandas>=2.0
reportlab>=4.0
pyyaml>=6.0
pytest>=7.0
